Refactor RenderUtils to speed it up
The StackWalker API is a LOT slower than using SecurityManager's getClassContext method. That, combined with using lighter checks to avoid even checking the calling class context (i.e. whether we are on the render thread at all and whether the item stack we're about to check all this for is even transmogged!)
Closes #1