Replace SecurityManager.getClassContext with mixin'd counters
The most performant way to check if you're in a class is just to have it tell you, as it turns out...
When other mods call vanilla methods that we mixin to, primarily Player.getItemBySlot, the previously harmless performance impact of SecurityManager.getClassContext spikes up dramatically...
The solution is to replace the checks for classes in the earlier stackframe with a counter that we increment in every single method in those classes (using mixins). This is *ugly*, and I don't like it at all, but it works and seems to have zero bugs while giving a huge performance boost, which is what's really important.